Support deserializing deeply nested sequences
It's possible for some Redis calls to return deeply nested sequneces
(eg. from a Lua script). This adds support for deserializing things with
arbitrary nesting.

Supporting arbitrary nesting resulted in a simplified design for
deserializer. To deserialize a nested sequence, we recurse into an
additional Deserializer. This contrasts with the previous design which
maintained additional state within a single deserializer. Furthermore,
no more Vec::reverse is needed; a peekable iterator is used instead.

To keep compatibility with the old API, a trait was added `IntoValueVec`
which is implemented for redis::Value and Vec<redis::Value>.