Version 0.5; ergonomic improvements
This fixes an API bug where creating a new deserializer returned a
Result that was always Ok. It now simply returns the deserializer
without possibility of error.

More importantly, this adds the RedisDeserialize trait which adds
`.deserialize()` to redis::Value to run the serde deserializer in a very
ergonomic way. See the updated readme for an example.